Move private data out of public repository
The path of your local git repository should not appear in the public
source code. For now, I have moved it to a private Class PrivateTestData:

package de.fau.osr;
final public class PrivateTestData {
    public static String getGitRepo() {
        return "*Your favorite path*/.git";
    }
}

This is better than the way it was before but not really a clean
solution. I do not think there is any clean way to run the unit tests on
the developement repository, so in the long term unit tests should be
run on some sample repository that could be placed in
src/test/resources/, which is addressable in a machine independant way
and gets copied once for every test run (so it is effectivly immutable).
Please voice your opinion on this!