Finished App for Req-5
This should in theory fulfill Req-5, however there is a bug in
GitVcsClient, where it tries to get the parent to a commit and it gets
the right commit id, but getTree() returns null! I have no idea why this
happens, it could be worked around by just instantiating a new RevWalk
to get each parent commit.