Track associated requirements with jgit
While looking at the jgit API I discovered that tracking which commit
created a line is far easier than I thought because jgit supports this
functionality with the BlameCommand / BlameGenerator classes.

This commit adds blame functionality to the GitVcsClient and returns a
list of lines with related commits. As the functionality for getting
aggregate linkage for a commitID is still not complete, I only use
information from the CommitMessageParser for now.