Make paths to files absolute, fixes #71
I introduce a getWorkingCopy method for the VcsClient that returns the
path of the current working copy (if there is any, returns garbage for
bare repositories right now). This is then used when checking for the
presence of files to get them right, independent of the current working
directory.
This pulls up the repository uri from GitVcsClient to VcsClient, since
it is not vcs specific.
The rationale for not making all paths absolute is that git needs paths
relative to the working copy for operations like blame and constructing
a composite path is far easier then deconstructing it.