Req-54, Req-55 Display additional info in eclipse
I extended getRequirementsLineLinkageForFile in Tracker to fetch and
format requirement details from database, that are then returned as a
List<Collection<String>> and put into multiple markers in eclipse. This
design was chosen, because for a single marker there is no obvious
possibility to display line breaks in the mouse over text.
For requirements that are not found in the database simply the id is
printed.