Req-70 Enable tracking modified lines by words
This commit enables the git backend of our software to track inserted and
deleted words instead of just lines. Also an extension point is already
implemented where custom information can be weaved into git historical
information.

The main class to enable this is GitBlameOperation, which right now computes
all information on the call to wordBlame(). As this is quite a slow operation,
the results should possibly be cached, and it would not be very hard to modify
wordBlame so that it provides a Stream of Annotations that can then be consumed
without locking up the whole GUI...
This is only the backend, if you want to know how stuff works, look at the test
class GitBlameOperationTest.

This commit also introduces the two utility classes
- FlatSource: stores a broken up string and the old line numbers
- AnnotatedWords: stores a FlatSource, together with Lists of annotations

I am not sure in what format to provide annotations, Object of course is not
very type-safe, on the other hand this flexibility means the blame does not have
to care about the req-commit relation anymore, like GitVcsClient.blame currently does.

Rename detection in the current state only detects perfect matches, but seeing
as detecting near matches is a complex problem that is not completely solveable
(just look at jgit / git sources), I would vote against implementing it at all
currently.

The TestRepository was updated to provide some a little more interesting data
for the wordBlame.