Improve blame performance #113
In GitBlameOperation, traverse the suspected lines and the edit list in
parallel, theoretically this gives a complexity of O(m+n) instead of
O(m*n) for this suboperation. Empirically, performance for the
wordBlame() operation improves by 10%-60% depending on the file.

The only drawback is that the code is more difficult to understand.