Fix #131:Tracibility Matrix progress bar not drawn
The progress bar was not drawn from the event processing thread, and the
progress value was not properly communicated across threads.

This commit moves all modifications of the progress bar to the event
thread, makes the progress of a RequirementsTraceabilityMatrixByImpact
an attribute, introduces an interface ProgressBarInterface to abstract
between backend and GUI, updates the progress bar via events instead of
polling, RequirementsTraceabilityMatrixByImpact now begins processing
only on request (to have a chance to attach the progressbar) and removes
the dead main from SpiceTraceabilityProgressBar.java. The type of the
window was changed to UTILITY, as POPUP potentially means "not
decorated, not movable".