Restructure FlatSource to conserve resources #113
FlatSource used to store a lot of String objects, which is very
expensive. It was modified to contain just a single byte buffer, that is
continuous and can be used directly for diff'ing.
Currently this saves approximately 5 seconds for a whole tracibility
matrix generation on our repository.

This commit also adds a test for wordsToLine, since the method was
changed, and it had no test previously.