Clean up TrackerAdapterWorker, #130
This commit reworks TrackerAdapterWorker so that is properly race free,
deadlock free and maintainable.

While after 0acac35 the previous code became nearly race free, it then
contained quite a few possibilities of deadlocks. While this could have
been solved with less changes, this rewrite hopefully makes this portion
of the codebase easier to read and modify.

While the TrackerAdapterWorker is now properly race free, the underlying
datastructures still are not, so it could still fail spontaneously!
(I think TracibilityMatrix generation is also affected by this?)

GenericLock was deleted as an unnecessary reimplementation of the
standard Lock class.

TrackerAdapterWorkerTest was deleted because it only tested methods that
were really uneeded in TrackerAdapterWorker. If we intend to use this
code, we should implement real tests for it.

A getHeadId method was added to VcsClient and all layers above it to be
able to efficiently determine the state of the vcs.