Fix #136: Req-32: Make CommitFiles in Commits lazy
In 117eef3, I mistakenly thought that the CommitFiles in the Commit
classes where never used, but failed to delete them both, so it silently
failed (#136). This commit makes commitFiles a lazy stream instead of
a Collection, because still in most cases commitFiles is not used, so a
lazy solution provides quite a performance boost.