Extend NOTES after looking at the code
Because it makes implementing Req-49 a lot harder, I just had a deeper
look at the DataSource structure. I tried to bring it into reasonable
shape, but ultimately failed (did not commit anything).

This commit adds a paragaph about that to the technical debt
documentation, since being extensible was a requirement of the industry
partner.